+++ rtlamr2mqtt.py
@@ -11,6 +11,11 @@ import paho.mqtt.client as mqtt
 
 # uses signal to shutdown and hard kill opened processes and self
 def shutdown(signum, frame):
+    mqtt_client.connect(host=mqtt_host, port=mqtt_port)
+    mqtt_client.loop_start()
+    mqtt_client.publish(topic=availability_topic, payload='offline', qos=0, retain=True)
+    mqtt_client.disconnect()
+    mqtt_client.loop_stop()
     if rtltcp.returncode is None:
         rtltcp.terminate()
         try:
@@ -55,31 +60,32 @@ with open('/etc/rtlamr2mqtt.yaml','r') as config_file:
   config = yaml.safe_load(config_file)
 
 # Build MQTT configuration
+availability_topic = 'rtlamr/status' # Setting for LWT messages
 mqtt_host = "127.0.0.1" if 'host' not in config['mqtt'] else config['mqtt']['host']
 mqtt_port = 1883 if 'port' not in config['mqtt'] else int(config['mqtt']['port'])
-mqtt_user = "user" if 'user' not in config['mqtt'] else config['mqtt']['user']
-mqtt_password = "secret" if 'password' not in config['mqtt'] else config['mqtt']['password']
+mqtt_user = None if 'user' not in config['mqtt'] else config['mqtt']['user']
+mqtt_password = None if 'password' not in config['mqtt'] else config['mqtt']['password']
 ha_autodiscovery_topic = "homeassistant" if 'ha_autodiscovery_topic' not in config['mqtt'] else str(config['mqtt']['ha_autodiscovery_topic'])
 ha_autodiscovery = False
 if 'ha_autodiscovery' in config['mqtt']:
     if str(config['mqtt']['ha_autodiscovery']).lower() in ['true', 'yes']:
         ha_autodiscovery = True
-
-state_topic = 'rtlamr/{}/state'
 mqtt_client = mqtt.Client(client_id='rtlamr2mqtt')
-mqtt_client.username_pw_set(username=mqtt_user, password=mqtt_password)
+if mqtt_user is not None:
+    mqtt_client.username_pw_set(username=mqtt_user, password=mqtt_password)
 
 if 'general' in config:
     sleep_for = 0 if 'sleep_for' not in config['general'] else config['general']['sleep_for']
 else:
     sleep_for = 0
 
-# Build RTLAMR config
+# Build Meter and RTLAMR config and send HA Auto-discover payload if enabled
 # TODO: Add a configuration section for rtlamr and rtl_tcp configuration parameters
 protocols = []
 meter_ids = []
 meter_readings = {}
 for idx,meter in enumerate(config['meters']):
+    state_topic = 'rtlamr/{}/state'.format(config['meters'][idx]['name'])
     config['meters'][idx]['name'] = str('meter_{}'.format(meter['id'])) if 'name' not in meter else str(meter['name'])
     config['meters'][idx]['unit_of_measurement'] = '' if 'unit_of_measurement' not in meter else str(meter['unit_of_measurement'])
     config['meters'][idx]['icon'] = 'mdi:gauge' if 'icon' not in meter else str(meter['icon'])
@@ -94,11 +100,15 @@ for idx,meter in enumerate(config['meters']):
             "name": config['meters'][idx]['name'],
             "unit_of_measurement": config['meters'][idx]['unit_of_measurement'],
             "icon": config['meters'][idx]['icon'],
-            "state_topic": state_topic.format(config['meters'][idx]['name'])
+            "availability_topic": availability_topic,
+            "state_topic": state_topic
         }
+        mqtt_client.will_set("stack/clientstatus", "offline", 1, True)
         mqtt_client.connect(host=mqtt_host, port=mqtt_port)
+        mqtt_client.loop_start()
         mqtt_client.publish(topic=discover_topic, payload=dumps(discover_payload), qos=0, retain=True)
         mqtt_client.disconnect()
+        mqtt_client.loop_stop()
 
 rtlamr_custom = []
 if 'custom_parameters' in config:
@@ -117,6 +127,11 @@ rtltcp_cmd = ['/usr/bin/rtl_tcp'] + rtltcp_custom
 
 # Main loop
 while True:
+    mqtt_client.connect(host=mqtt_host, port=mqtt_port)
+    mqtt_client.loop_start()
+    mqtt_client.publish(topic=availability_topic, payload='online', qos=0, retain=True)
+    mqtt_client.disconnect()
+    mqtt_client.loop_stop()
     # Is this the first time are we executing this loop? Or is rtltcp running?
     if 'rtltcp' not in locals() or rtltcp.poll() is not None:
         # start the rtl_tcp program
@@ -153,9 +168,12 @@ while True:
                         else:
                             formated_reading = raw_reading
                         print('Meter "{}" - Consumption {}. Sending value to MQTT.'.format(meter_id, formated_reading), file=sys.stderr)
+                        state_topic = 'rtlamr/{}/state'.format(meter['name'])
+                        mqtt_client.loop_start()
                         mqtt_client.connect(host=mqtt_host, port=mqtt_port)
-                        mqtt_client.publish(topic=state_topic.format(meter['name']), payload=str(formated_reading).encode('utf-8'), qos=0, retain=True)
+                        mqtt_client.publish(topic=state_topic, payload=str(formated_reading).encode('utf-8'), qos=0, retain=True)
                         mqtt_client.disconnect()
+                        mqtt_client.loop_stop()
                         meter_readings[meter_id] += 1
         if sleep_for > 0:
             # Check if we have readings for all meters