+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ### 2021-09-09
+  - Added last Will and testment messages
+  - Added availability status topic
   - Added RTL_MSGTYPE to debug mode
 
 ### 2021-09-03